Overhauled the sequencer features to make it generate the required amount of points, and update example [STILL IN DEBUG FORM]

+++ src/openafe.h
@@ -163,7 +163,7 @@ class AFE
 		 * @return true if there is data available
 		 * @return false if there is no data available
 		 */
-		static uint32_t dataAvailable(void);
+		static uint16_t dataAvailable(void);
 
 		/**
 		 * @brief Start the voltametry.
